security-toolkit/tool-05-jwt-analyzer/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
cryptography==41.0.7
requests==2.31.0
reportlab==4.0.7
